+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+* Support the legacy React context API, fixing [#7](https://github.com/jaydenseric/graphql-react/issues/7).
 * Use [`eslint-plugin-ava`](https://npm.im/eslint-plugin-ava).
 
 ## 1.0.0-alpha.2